Se agregan las constantes en el proyecto
