--- Infrastructure/Api/ApiEndpoints.swift
@@ -1,15 +1,0 @@
-//
-//  Configuration.swift
-//  PruebaCeiba
-//
-//  Created by Sergio Josue Rodriguez Taborda on 6/12/22.
-//
-
-import Foundation
-
-class ApiEndpoints
-{
-    let apiKey : String? = Bundle.main.infoDictionary?["URL_BASE"] as? String
-}
-
-
