+++ Podfile
@@ -1,0 +1,14 @@
+# Uncomment the next line to define a global platform for your project
+# platform :ios, '9.0'
+
+target 'PruebaCeiba' do
+  # Comment the next line if you don't want to use dynamic frameworks
+  use_frameworks!
+
+pod 'Alamofire'
+pod "Factory"
+pod 'RealmSwift'
+
+  # Pods for PruebaCeiba
+
+end
