--- Presentation/Model/UserPublicationModel.swift
@@ -1,13 +1,0 @@
-//
-//  UserPublication.swift
-//  PruebaCeiba
-//
-//  Created by Sergio Josue Rodriguez Taborda on 7/12/22.
-//
-
-import Foundation
-
-struct UserPublicationModel: Identifiable {
-    let userID, id: Int
-    let title, body: String
-}
