--- Presentation/View/ProgressIndicator/ProgressIndicator.swift
@@ -1,22 +1,0 @@
-//
-//  ProgressIndicator.swift
-//  PruebaCeiba
-//
-//  Created by Sergio Josue Rodriguez Taborda on 7/12/22.
-//
-
-import SwiftUI
-
-struct ProgressIndicator: View {
-    var body: some View {
-        ProgressView()
-            .progressViewStyle(CircularProgressViewStyle(tint: .gray))
-            .scaleEffect(3)
-    }
-}
-
-struct ProgressIndicator_Previews: PreviewProvider {
-    static var previews: some View {
-        ProgressIndicator()
-    }
-}
