--- Presentation/View/PruebaCeibaApp.swift
@@ -1,20 +1,0 @@
-//
-//  PruebaCeibaApp.swift
-//  PruebaCeiba
-//
-//  Created by Sergio Josue Rodriguez Taborda on 5/12/22.
-//
-
-import SwiftUI
-
-@available(iOS 16.0, *)
-@main
-struct PruebaCeibaApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-            //UserPublicationCard()
-           // UserDetail()
-        }
-    }
-}
