--- Presentation/View/UserList.swift
@@ -1,20 +1,0 @@
-//
-//  UserList.swift
-//  PruebaCeiba
-//
-//  Created by Sergio Josue Rodriguez Taborda on 5/12/22.
-//
-
-import SwiftUI
-
-struct UserList: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct UserList_Previews: PreviewProvider {
-    static var previews: some View {
-        UserList()
-    }
-}
