+++ PruebaCeiba/Infrastructure/Api/ApiEndpoints.swift
@@ -1,0 +1,17 @@
+//
+//  Configuration.swift
+//  PruebaCeiba
+//
+//  Created by Sergio Josue Rodriguez Taborda on 6/12/22.
+//
+
+import Foundation
+
+class ApiEndpoints
+{
+    let urlBase = "https://jsonplaceholder.typicode.com"
+    let getUsers = "/users"
+    let getPublicationsByIdUser = "/posts?userId="
+}
+
+
