+++ PruebaCeiba/Infrastructure/Database/ErrorDatabase.swift
