+++ PruebaCeiba/Presentation/View/ProgressIndicator/ProgressIndicator.swift
@@ -1,0 +1,24 @@
+//
+//  ProgressIndicator.swift
+//  PruebaCeiba
+//
+//  Created by Sergio Josue Rodriguez Taborda on 7/12/22.
+//
+
+import SwiftUI
+
+struct ProgressIndicator: View {
+    var body: some View {
+        ProgressView()
+            .progressViewStyle(CircularProgressViewStyle(tint: .gray))
+            .scaleEffect(3)
+    }
+}
+
+#if DEBUG
+struct ProgressIndicator_Previews: PreviewProvider {
+    static var previews: some View {
+        ProgressIndicator()
+    }
+}
+#endif
