+++ PruebaCeiba/Presentation/View/PruebaCeibaApp.swift
@@ -1,0 +1,17 @@
+//
+//  PruebaCeibaApp.swift
+//  PruebaCeiba
+//
+//  Created by Sergio Josue Rodriguez Taborda on 5/12/22.
+//
+
+import SwiftUI
+
+@main
+struct PruebaCeibaApp: App {
+    var body: some Scene {
+        WindowGroup {
+            UserList()
+        }
+    }
+}
