+++ PruebaCeiba/Utils/Constants.swift
@@ -1,0 +1,35 @@
+//
+//  Constants.swift
+//  PruebaCeiba
+//
+//  Created by Sergio Josue Rodriguez Taborda on 8/12/22.
+//
+
+import Foundation
+
+struct Constants
+{
+    // MARK: - Colors
+    static let colorGreenPrimary = "GreenPrimary"
+    static let colorBackground = "BackGroundColor"
+    
+    // MARK: - General
+    static let titleBarApp = "Prueba de Ingreso"
+    static let apiUrlNotFound = "Url no valida"
+    
+    // MARK: - UserCard
+    static let cardUserVerPublicaciones = "VER PUBLICACIONES"
+    static let cardUserIconPhone = "phone.fill"
+    static let cardUserIconEmail = "envelope.fill"
+    
+    // MARK: - ListIsEmpty
+    static let listIsEmpty = "List is empty"
+    
+    
+    // MARK: - UserPublicationCard
+    static let userPublicationCardId = "Id:"
+    static let userPublicationCardUserId = "UserId:"
+    
+    // MARK: - UserList
+    static let userListSearchPlaceholder = "Buscar usuario"
+}
