+++ PruebaCeibaUITests/PruebaCeibaUITestsLaunchTests.swift
@@ -1,0 +1,32 @@
+//
+//  PruebaCeibaUITestsLaunchTests.swift
+//  PruebaCeibaUITests
+//
+//  Created by Sergio Josue Rodriguez Taborda on 9/12/22.
+//
+
+import XCTest
+
+final class PruebaCeibaUITestsLaunchTests: XCTestCase {
+
+    override class var runsForEachTargetApplicationUIConfiguration: Bool {
+        true
+    }
+
+    override func setUpWithError() throws {
+        continueAfterFailure = false
+    }
+
+    func testLaunch() throws {
+        let app = XCUIApplication()
+        app.launch()
+
+        // Insert steps here to perform after app launch but before taking a screenshot,
+        // such as logging into a test account or navigating somewhere in the app
+
+        let attachment = XCTAttachment(screenshot: app.screenshot())
+        attachment.name = "Launch Screen"
+        attachment.lifetime = .keepAlways
+        add(attachment)
+    }
+}
